requirements.txt
aws-cdk-lib==2.155.0
constructs>=10.0.0,<11.0.0
boto3
fastapi
llama-index  
llama-index-embeddings-openai
llama-index-vector-stores-qdrant
llama-index-readers-s3
qdrant-client
python-dotenv
uvicorn
openai
pytest
